# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InviteCreateRequest": {
        "additionalProperties": false,
        "description": "The delivery address and authority an owner or admin selected.",
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+$",
            "title": "Email",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          }
        },
        "required": [
          "email",
          "role"
        ],
        "title": "InviteCreateRequest",
        "type": "object"
      },
      "InviteResponse": {
        "additionalProperties": false,
        "description": "One invitation including the bearer token disclosed only on creation.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expiresAt": {
            "title": "Expiresat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "token",
          "expiresAt"
        ],
        "title": "InviteResponse",
        "type": "object"
      },
...
      "MemberCollectionResponse": {
        "additionalProperties": false,
        "description": "The people who can currently act inside one Workspace.",
        "properties": {
          "members": {
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            },
            "title": "Members",
            "type": "array"
          }
        },
        "required": [
          "members"
        ],
        "title": "MemberCollectionResponse",
        "type": "object"
      },
      "MemberResponse": {
        "additionalProperties": false,
        "description": "One fellow member as it is safe to show inside their Workspace.",
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "joinedAt": {
            "title": "Joinedat",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          },
          "userId": {
            "format": "uuid",
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "role",
          "displayName",
          "email",
          "joinedAt"
        ],
        "title": "MemberResponse",
        "type": "object"
      },
      "MemberRoleRequest": {
        "additionalProperties": false,
        "description": "The explicit non-owner role selected for an active member.",
        "properties": {
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          }
        },
        "required": [
          "role"
        ],
        "title": "MemberRoleRequest",
        "type": "object"
      },
...
      "OwnershipTransferRequest": {
        "additionalProperties": false,
        "description": "The active member who will become the Workspace owner.",
        "properties": {
          "userId": {
            "format": "uuid",
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId"
        ],
        "title": "OwnershipTransferRequest",
        "type": "object"
      },
      "PendingInviteCollectionResponse": {
        "additionalProperties": false,
        "description": "Every invitation that remains available for acceptance.",
        "properties": {
          "invites": {
            "items": {
              "$ref": "#/components/schemas/PendingInviteResponse"
            },
            "title": "Invites",
            "type": "array"
          }
        },
        "required": [
          "invites"
        ],
        "title": "PendingInviteCollectionResponse",
        "type": "object"
      },
...
      "PublishingRolePolicy": {
        "enum": [
          "owner_admin_editor",
          "owner_admin"
        ],
        "title": "PublishingRolePolicy",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkspaceCollectionResponse": {
        "additionalProperties": false,
        "description": "Every Workspace the caller may enter.",
        "properties": {
          "workspaces": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceResponse"
            },
            "title": "Workspaces",
            "type": "array"
          }
        },
        "required": [
          "workspaces"
        ],
        "title": "WorkspaceCollectionResponse",
        "type": "object"
      },
      "WorkspaceCreateRequest": {
        "additionalProperties": false,
        "description": "The Workspace a member asks us to create for them.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/WorkspaceKind",
            "default": "team"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "WorkspaceCreateRequest",
        "type": "object"
      },
      "WorkspaceKind": {
        "enum": [
          "personal",
          "team"
        ],
        "title": "WorkspaceKind",
        "type": "string"
      },
      "WorkspaceResponse": {
        "additionalProperties": false,
        "description": "One Workspace as it is safe to show a member of it.",
        "properties": {
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/WorkspaceKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "publishingRolePolicy": {
            "$ref": "#/components/schemas/PublishingRolePolicy"
          },
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkspaceStatus"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "kind",
          "status",
          "publishingRolePolicy",
          "role",
          "createdAt"
        ],
        "title": "WorkspaceResponse",
        "type": "object"
      },
      "WorkspaceRole": {
        "enum": [
          "owner",
          "admin",
          "editor",
          "reviewer",
          "viewer"
        ],
        "title": "WorkspaceRole",
        "type": "string"
      },
      "WorkspaceStatus": {
        "enum": [
          "active",
          "suspended",
          "deleted"
        ],
        "title": "WorkspaceStatus",
        "type": "string"
      },
...
      "WorkspaceUpdateRequest": {
        "additionalProperties": false,
        "description": "The Workspace settings an owner or admin may change.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "publishingRolePolicy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublishingRolePolicy"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "WorkspaceUpdateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Clipah API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/workspace-invites/{token}/accept": {
      "post": {
        "description": "Accept one invitation as the independently authenticated current User.",
        "operationId": "accept_workspace_invite_api_v1_workspace_invites__token__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Workspace Invite",
        "tags": [
          "workspace-memberships"
        ]
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "description": "List every Workspace the caller may enter.",
        "operationId": "index_api_v1_workspaces_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Index",
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "description": "Create a Workspace owned by the caller.",
        "operationId": "create_api_v1_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}": {
      "get": {
        "description": "Describe one Workspace to one of its members.",
        "operationId": "show_api_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Show",
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "description": "Change the Workspace settings its stewards control.",
        "operationId": "update_api_v1_workspaces__workspace_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/invites": {
      "get": {
        "description": "List active invitation metadata without disclosing bearer tokens.",
        "operationId": "list_workspace_invites_api_v1_workspaces__workspace_id__invites_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingInviteCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workspace Invites",
        "tags": [
          "workspace-memberships"
        ]
      },
      "post": {
        "description": "Create one role-bound team Workspace invitation.",
        "operationId": "create_workspace_invite_api_v1_workspaces__workspace_id__invites_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workspace Invite",
        "tags": [
          "workspace-memberships"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/invites/{invite_id}": {
      "delete": {
        "description": "Revoke one invitation before it is accepted.",
        "operationId": "revoke_workspace_invite_api_v1_workspaces__workspace_id__invites__invite_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "invite_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invite Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Workspace Invite",
        "tags": [
          "workspace-memberships"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/members": {
      "get": {
        "description": "List the people who can currently act inside this Workspace.",
        "operationId": "members_api_v1_workspaces__workspace_id__members_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Members",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/members/{member_user_id}": {
      "delete": {
        "description": "Remove one member and make every existing Session lose tenant access.",
        "operationId": "remove_workspace_member_api_v1_workspaces__workspace_id__members__member_user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "member_user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Member User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Workspace Member",
        "tags": [
          "workspace-memberships"
        ]
      },
      "patch": {
        "description": "Assign one explicit non-owner role to an active member.",
        "operationId": "update_workspace_member_api_v1_workspaces__workspace_id__members__member_user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "member_user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Member User Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Workspace Member",
        "tags": [
          "workspace-memberships"
        ]
      }
    },
    "/api/v1/workspaces/{workspace_id}/ownership-transfers": {
      "post": {
        "description": "Transfer team Workspace ownership to one active member atomically.",
        "operationId": "transfer_workspace_ownership_api_v1_workspaces__workspace_id__ownership_transfers_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfer Workspace Ownership",
        "tags": [
          "workspace-memberships"
        ]
      }
    },
...
  }
}